requirements.txt
requests
pandas
python-dotenv
streamlit
plotly-express